pyproject.toml
[project]
name = "tabi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.57.0",
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
]
